Module Start.App_result

The result of an application-level component is a value of type 'incoming App_result.t. This value contains the view of the app, and also an inject function: a way for the holder of the Handle.t to send events into the application component.

If the application developer doesn't want to use incoming events, they should use Core_kernel.Nothing.t for the 'incoming type, and Core_kernel.Nothing.unreachable_code for the value of inject_incoming.

type ('extra, 'incoming) t

val create : view:Bonsai_revery__.Import.Element.t -> extra:'extra -> inject_incoming:('incoming -> Bonsai_revery__.Import.Event.t) -> ('extra, 'incoming) t