Module Bonsai_revery__.Start

module Handle : sig ... end

module App_input : sig ... end

The input of an application-level component is a value of type ('input, 'outgoing) App_input.t.

module App_result : sig ... end

The result of an application-level component is a value of type 'incoming App_result.t. This value contains the view of the app, and also an inject function: a way for the holder of the Handle.t to send events into the application component.

Start functions

val start_standalone : initial_input:'input -> ('input, Bonsai_revery__.Import.Element.t) Bonsai_revery__.Import.Bonsai.t -> unit

Start an application, receiving a handle that can't schedule any actions.

The outermost Bonsai.t's result type parameter should be an Element.t, which will be rendered inside a window.

val start : initial_input:'input -> ?⁠with_handle:(('input, 'extra, 'incoming, 'outgoing) Handle.t -> unit Lwt.t) -> (('input, 'outgoing) App_input.t, ('extra, 'incoming) App_result.t) Bonsai_revery__.Import.Bonsai.t -> unit

Start an application, receiving a handle that can schedule actions of a user-defined type.

The outermost Bonsai.t's result type parameter should be a pair consisting of:

an Element.t, which will be rendered inside a window; and

an inject function that accepts external actions and returns Event.ts.