Module Start.App_input

The input of an application-level component is a value of type ('input, 'outgoing) App_input.t.

type ('input, 'outgoing) t

val input : ('input, _) t -> 'input

The input field is used to access the 'input to the application that is set from the imperative Handle.t value.

val inject_outgoing : (_, 'outgoing) t -> 'outgoing -> Bonsai_revery__.Import.Event.t

inject_outgoing is used to inject values of the type 'outgoing into events to communicate with the imperative Handle.t holder. Any values injected via inject_outgoing will be present in the 'outgoing Pipe.Reader.t that can be acquired via Handle.outgoing.